requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
ansi2html==1.8.0
anyio==3.6.2
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
azure-functions==1.14.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==2.1.1
click==8.1.3
cmdstanpy==1.1.0
colorama==0.4.6
contourpy==1.0.7
convertdate==2.4.0
cycler==0.11.0
Cython==3.0.0
dash==2.11.1
dash-bootstrap-components==1.4.1
dash-bootstrap-templates==1.0.8
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
DatastreamDSWS==1.0.10
DateTime==5.1
dnspython==2.3.0
ephem==4.1.4
Flask==2.2.5
fonttools==4.39.3
frozenlist==1.4.0
greenlet==2.0.2
gunicorn==21.2.0
h11==0.12.0
holidays==0.29
httpcore==0.15.0
httpx==0.23.0
idna==3.4
importlib-resources==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
LunarCalendar==0.0.9
MarkupSafe==2.1.2
matplotlib==3.7.1
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.23.4
openai==0.27.8
packaging==23.1
pandas==1.5.3
patsy==0.5.3
Pillow==10.0.0
plotly==5.14.1
pmdarima==2.0.3
prophet==1.1.4
pyee==9.0.4
pyhumps==3.0.2
PyMeeus==0.5.12
pymongo==4.3.3
PyMySQL==1.0.3
pyodbc
pyparsing==3.0.9
python-dateutil==2.8.2
python-stdnum==1.18
pytz==2023.3
refinitiv-data==1.3.0
requests==2.28.1
retrying==1.3.4
rfc3986==1.5.0
scikit-learn==1.3.0
scipy==1.9.3
simplejson==3.19.1
six==1.16.0
scikit-learn
sniffio==1.3.0
SQLAlchemy==2.0.12
statsmodels==0.14.0
tenacity==8.0.1
threadpoolctl==3.2.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
waitress==2.1.2
watchdog==2.1.9
websocket-client==1.4.1
Werkzeug==2.2.3
wincertstore==0.2
yarl==1.9.2
zope.interface==6.0